pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["romario lima <romariolima1006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
factory-boy = "^3.3.0"
django = "^5.0"
pytest = "^7.4.3"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
django-test = "^0.4030"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
